Refactor backtest project to use the new App.py
This commit basically replace the old App.py to the new App2.py.
Also renamed App2.py to be App.py.
Also deleted the old "params.json" since we will now use the new JSON files.

I believe we are now in a good spot to implement new features on Backtest.
