--- main/backtest/resources/params.json
@@ -1,54 +1,0 @@
-{
-    "pair": "BTCUSDT",
-    "startTime": "01.08.2024",
-    "endTime": "17.10.2024",
-    "period_label": "bear",
-    "period": "1h",
-    "trend_period": "D",
-    "_comment": "trend period possibilities: 'W', 'W-SUN', 'D', 'h', '4h', '2h30min' ", 
-    "filter_buy_classes": [
-        "FilterBuy_RSI",
-        "FilterBuy_RSI_price_x_SMAlong",
-        "FilterBuy_RSI_price_x_SMAmedium",
-        "FilterBuy_RSI_EMAshort_gt_SMAlong_price_x_SMAlong",
-        "FilterBuy_RSI_EMAshort_gt_SMAlong_price_x_SMAmedium",
-        "FilterBuy_RSI_EMAshort_gt_SMAlong",
-        "FilterBuy_EMAshort_lt_SMAmedium_gt_SMAlong_Price_gt_SMAmedium"
-    ],
-    "trigger_buy_classes": [
-        "TriggeredState_MaxCandles"
-    ],
-    "trade_buy_classes": [
-        "TradeBuy_High_x_HighLastCandle",
-        "TradeBuy_HighLastCandle",
-        "TradeBuy_Price_gt_EMAshort",
-        "TradeBuy_Price_gt_SMAmedium",
-        "TradeBuy_EMAshort_gt_SMAmedium",
-        "TradeBuy_EMAshort_gt_SMAmedium_High_gt_HighLastCandle",
-        "TradeBuy_Close_gt_CloseLastCandle",
-        "TradeBuy_HighLastCandle_EMAshort_gt_SMAmedium"
-    ],
-    "filter_sell_classes": [
-        "FilterSell_RSI",
-        "FilterSell_alwaysTrue"
-    ],
-    "trigger_sell_classes": [
-        "TriggeredState_MaxCandles_Sell"
-    ],
-    "trade_sell_classes": [
-        "TradeSell_LowLastCandle",
-        "TradeSell_Price_EMAshort"
-    ],
-    "trend_classes": [
-        "UpTrend_AlwaysTrend",
-        "UpTrend_EMAshort_gt_SMAlong",
-        "UpTrend_EMAshort_gt_SMAmedium",
-        "UpTrend_SMAmedium_gt_SMAlong",
-        "UpTrend_EMAshort_gt_SMAmedium_gt_SMAlong",
-        "UpTrend_Price_gt_SMAmedium",
-        "UpTrend_Price_gt_SMAlong"
-    ],
-    "strategy_classes": [
-        "B2"
-    ]
-}
